feat: lib/bin split
